+++ tests/fpga/streaming_memory_test.py
@@ -12,9 +12,17 @@ from dace.fpga_testing import xilinx_test
 
 M, N, K = 64, 64, 64
 
+
 @dace.program
 def vecadd_1_streaming(A: dace.float32[N], B: dace.float32[N]):
     B[:] = A + 1.0
+
+
+@dace.program
+def vecadd_streaming(A: dace.float32[N], B: dace.float32[N],
+                     C: dace.float32[N]):
+    C[:] = A + B
+
 
 @dace.program
 def matadd_streaming(A: dace.float32[M, N], B: dace.float32[M, N],
@@ -27,6 +35,24 @@ def matadd_multistream(A: dace.float32[M, N], B: dace.float32[M, N],
                        C: dace.float32[M, N], D: dace.float32[M, N]):
     C[:] = A + B
     D[:] = A - B
+
+@dace.program
+def matmul_streaming(A: dace.float32[M, K], B: dace.float32[K, N],
+                     C: dace.float32[M, N]):
+    tmp = np.ndarray([M, N, K], dtype=A.dtype)
+
+    # Multiply every pair of values to a large 3D temporary array
+    for i, j, k in dace.map[0:M, 0:N, 0:K]:
+        with dace.tasklet:
+            in_A << A[i, k]
+            in_B << B[k, j]
+            out >> tmp[i, j, k]
+
+            out = in_A * in_B
+
+    # Sum last dimension of temporary array to obtain resulting matrix
+    dace.reduce(lambda a, b: a + b, tmp, C, axis=2, identity=0)
+
 
 
 @dace.program
@@ -239,6 +265,7 @@ def test_streaming_and_composition():
 
     return sdfg
 
+
 @xilinx_test()
 def test_mem_buffer_vec_add_1():
     # Make SDFG
@@ -250,10 +277,13 @@ def test_mem_buffer_vec_add_1():
         InlineSDFG,
     ])
 
-    # sdfg.apply_transformations_repeated(mb.MemoryBuffering)
-
-    # assert sdfg.apply_transformations_repeated(
-    #     mb.MemoryBuffering, dict(storage=dace.StorageType.FPGA_Local)) == 3
+    assert sdfg.apply_transformations_repeated(sm.StreamingMemory,
+                                        options=[{
+                                            'use_memory_buffering':
+                                            True,
+                                            "storage":
+                                            dace.StorageType.FPGA_Local
+                                        }]) == 2
 
     # Run verification
     A = np.random.rand(N).astype(np.float32)
@@ -262,6 +292,98 @@ def test_mem_buffer_vec_add_1():
     sdfg(A=A, B=B)
 
     assert all(B == A + 1)
+
+    return sdfg
+
+
+@xilinx_test()
+def test_mem_buffer_vec_add():
+    # Make SDFG
+    sdfg: dace.SDFG = vecadd_streaming.to_sdfg()
+    # Transform
+
+    sdfg.apply_transformations([
+        FPGATransformSDFG,
+        InlineSDFG,
+    ])
+
+    assert sdfg.apply_transformations_repeated(sm.StreamingMemory,
+                                        options=[{
+                                            'use_memory_buffering':
+                                            True,
+                                            "storage":
+                                            dace.StorageType.FPGA_Local
+                                        }]) == 3
+
+    # Run verification
+    A = np.random.rand(N).astype(np.float32)
+    B = np.random.rand(N).astype(np.float32)
+    C = np.random.rand(N).astype(np.float32)
+
+    sdfg(A=A, B=B, C=C)
+
+    diff = np.linalg.norm(C - (A + B))
+    assert diff <= 1e-5
+
+    return sdfg
+
+
+@xilinx_test()
+def test_mem_buffer_mat_add():
+    # Make SDFG
+    sdfg: dace.SDFG = matadd_streaming.to_sdfg()
+    # Transform
+    sdfg.apply_transformations([FPGATransformSDFG, InlineSDFG])
+
+    assert sdfg.apply_transformations_repeated(sm.StreamingMemory,
+                                        options=[{
+                                            'use_memory_buffering':
+                                            True,
+                                            "storage":
+                                            dace.StorageType.FPGA_Local
+                                        }]) == 3
+
+    # Run verification
+    A = np.random.rand(M, N).astype(np.float32)
+    B = np.random.rand(M, N).astype(np.float32)
+    C = np.random.rand(M, N).astype(np.float32)
+
+    sdfg(A=A, B=B, C=C)
+
+    diff = np.linalg.norm(C - (A + B))
+
+    assert diff <= 1e-5
+
+    return sdfg
+
+
+@xilinx_test()
+def test_mem_buffer_mat_mul():
+    # Make SDFG
+    sdfg: dace.SDFG = matmul_streaming.to_sdfg()
+    # Transform
+    sdfg.apply_transformations([FPGATransformSDFG, InlineSDFG])
+
+    sdfg.apply_transformations(sm.StreamingMemory,
+                                        options=[{
+                                            'use_memory_buffering':
+                                            True,
+                                            "storage":
+                                            dace.StorageType.FPGA_Local
+                                        }])
+
+    # assert sdfg.apply_transformations_repeated(
+    #     mb.MemoryBuffering, dict(storage=dace.StorageType.FPGA_Local)) == 3
+
+    # Run verification
+    A = np.random.rand(M, K).astype(np.float32)
+    B = np.random.rand(K, N).astype(np.float32)
+    C = np.random.rand(M, N).astype(np.float32)
+
+    sdfg(A=A, B=B, C=C)
+
+    diff = np.linalg.norm(C - (A @ B))
+    assert diff <= 1e-5
 
     return sdfg
 
@@ -278,3 +400,9 @@ if __name__ == "__main__":
     # test_streaming_and_composition(None)
 
     test_mem_buffer_vec_add_1(None)
+    test_mem_buffer_vec_add(None)
+    test_mem_buffer_mat_add(None)
+    # test_mem_buffer_mat_mul(None)
+
+    # TODO: Write more test cases
+    # Symbol
